requirements.txt
streamlit==1.27.0
pandas==1.5.3
numpy==1.24.3
pandasai==0.7.0
openai==0.27.10
xlrd==2.0.1
openpyxl==3.1.2
python-dotenv